# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  config.FlareSolverrConfig:
    properties:
      enabled:
        description: Enable FlareSolverr for bypassing Cloudflare
        type: boolean
      max_retries:
        description: 'Max retry attempts for FlareSolverr calls (default: 3)'
        type: integer
      session_ttl:
        description: 'Session TTL in seconds (default: 300)'
        type: integer
      timeout:
        description: 'Request timeout in seconds (default: 30)'
        type: integer
      url:
        description: 'FlareSolverr endpoint (default: http://localhost:8191/v1)'
        type: string
    type: object
...
  config.ProxyConfig:
    properties:
      default_profile:
        description: Default profile name (for global scrapers.proxy)
        type: string
      enabled:
        description: Enable proxy for HTTP requests
        type: boolean
      profile:
        description: Named profile to use (for scraper-specific overrides)
        type: string
      profiles:
        additionalProperties:
          $ref: '#/definitions/config.ProxyProfile'
        description: Named proxy profiles (global scrapers.proxy)
        type: object
    type: object
  config.ProxyProfile:
    properties:
      password:
        type: string
      url:
        type: string
      username:
        type: string
    type: object
...
  system.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  system.ProxyTestRequest:
    properties:
      flaresolverr:
        allOf:
        - $ref: '#/definitions/config.FlareSolverrConfig'
        description: FlareSolverr config (separate from ProxyConfig)
      mode:
        description: direct or flaresolverr
        enum:
        - direct
        - flaresolverr
        type: string
      proxy:
        $ref: '#/definitions/config.ProxyConfig'
      target_url:
        description: Optional override target URL
        type: string
    required:
    - mode
    type: object
  system.ProxyTestResponse:
    properties:
      duration_ms:
        type: integer
      flaresolverr_url:
        description: FlareSolverr endpoint used
        type: string
      message:
        type: string
      mode:
        type: string
      proxy_url:
        description: Redacted proxy URL
        type: string
      status_code:
        type: integer
      success:
        type: boolean
      target_url:
        type: string
    type: object
...
  temp.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/posters/{filename}:
    get:
      description: Serves persistent cropped posters from the database. These persist
        across scraping sessions.
      parameters:
      - description: Filename
        in: path
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/temp.ErrorResponse'
      summary: Serve cropped poster image
  /api/v1/proxy/test:
    post:
      consumes:
      - application/json
      description: Test direct proxy or FlareSolverr access to a target URL using
        provided proxy settings
      parameters:
      - description: Proxy test request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.ProxyTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.ProxyTestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/system.ErrorResponse'
      summary: Test proxy connectivity
      tags:
      - system
...
  /api/v1/temp/image:
    get:
      description: Proxies remote images for preview UI, handling hotlink protection
        and CORS issues.
      parameters:
      - description: Image URL
        in: query
        name: url
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/temp.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/temp.ErrorResponse'
      summary: Proxy remote images
  /api/v1/temp/posters/{jobId}/{filename}:
    get:
      description: Serves temporarily cropped posters from batch jobs. These are ephemeral
        and preserved when organization fails for retry.
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Filename
        in: path
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/temp.ErrorResponse'
      summary: Serve temporary poster image
...
schemes:
- http
- https
swagger: "2.0"